% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethiopian_to_greg.R
\name{ethiopian_to_greg}
\alias{ethiopian_to_greg}
\title{Ethiopian calendar to Gregorian Calendar (in CMC format) Conversion}
\usage{
ethiopian_to_greg(cmc = NULL)
}
\arguments{
\item{cmc}{Ethiopian Century Month Code defaults to NULL}
}
\description{
Ethiopian calendar to Gregorian Calendar (in CMC format) Conversion
}
\examples{
ethiopian_to_greg()
}
\author{
Heeju Sohn
}
\keyword{Calendar}
\keyword{DHS}
\keyword{Ethiopian}
