% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nepal_to_greg.R
\name{nepal_to_greg}
\alias{nepal_to_greg}
\title{Nepalese calendar to Gregorian Calendar (in CMC format) Conversion}
\usage{
nepal_to_greg(cmc = NULL)
}
\arguments{
\item{cmc}{Nepalese Century Month Code defaults to NULL}
}
\description{
Nepalese calendar to Gregorian Calendar (in CMC format) Conversion
}
\examples{
nepal_to_greg()
}
\author{
Heeju Sohn
}
\keyword{Calendar}
\keyword{DHS}
\keyword{Nepalese}
