% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persian_to_greg.R
\name{persian_to_greg}
\alias{persian_to_greg}
\title{Persian calendar to Gregorian Calendar (in CMC format) Conversion}
\usage{
persian_to_greg(ydhs = NULL, mdhs = NULL, cmc = NULL)
}
\arguments{
\item{ydhs}{Persian year defaults to NULL}

\item{mdhs}{Persian month defaults to NULL}

\item{cmc}{Persian Century Month Code defaults to NULL}
}
\description{
Persian calendar to Gregorian Calendar (in CMC format) Conversion
}
\examples{
persian_to_greg()
}
\author{
Heeju Sohn
}
\keyword{Calendar}
\keyword{DHS}
\keyword{Persian}
