% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thai_to_greg.R
\name{thai_to_greg}
\alias{thai_to_greg}
\title{Thai calendar to Gregorian Calendar (in CMC format) Conversion}
\usage{
thai_to_greg(cmc = NULL)
}
\arguments{
\item{cmc}{Thai Century Month Code defaults to NULL}
}
\description{
Thai calendar to Gregorian Calendar (in CMC format) Conversion
}
\examples{
thai_to_greg()
}
\author{
Heeju Sohn
}
\keyword{Calendar}
\keyword{DHS}
\keyword{Thai}
